requirements.txt
python-dotenv
google-generativeai
pytesseract
Pillow
pdf2image
python-docx
openpyxl
